requirements.txt
asgiref==3.10.0
dj-database-url==0.5.0
Django==4.2.26
psycopg2-binary==2.9.11
sqlparse==0.5.3
typing_extensions==4.15.0
django-crispy-forms
crispy-bootstrap5
